+++ data/monthly_sales.sh
@@ -1,1 +1,3 @@
 CQChartsTest -tcl -exec monthly_sales.tcl -close_app
+
+#CQChartsTest -script -exec monthly_sales.script -close_app