+++ data/process_model.tcl
@@ -7,8 +7,8 @@ set nr [get_property -model $modelId -name num_rows]
 set rc [get_property -model $modelId -name num_columns]
 
 #process_model -ind $modelId -add "column(2)/20000.0"
-process_model -ind $modelId -add "\[remap 2 0 36\]" -header "symbol size"
-process_model -ind $modelId -add "\[remap 2 0 1\]" -header color
+process_model -ind $modelId -add "remap(2,0,36)" -header "symbol size"
+process_model -ind $modelId -add "remap(2,0,1)" -header color
 
 # columns x, y, name symbolSize, fontSize, color, id
 #remove_plot -view view1 -all